[issue-11816] add support for im.vector.setting.override_colors
im.vector.setting.override_colors is a map of userId -> colorSpec

colorSpec is either a color index (5) or a hexadecimal color (#rrggbb)

If colorSpec is a color index then this index will be used instead of
the userId hash to calculate the color-class for the nick-name span.

If colorSpec is a hexadecimal color code, then the span color will be
specified with a style.color attribute.

Setting the override color is currently only possible from the android
client. see https://github.com/vector-im/element-android/pull/2614